rebase onto main and fix broken tests
